feat: Certification Authorization

+++ src/main/java/com/daniel/community/post/service/PostService.java
@@ -1,45 +1,50 @@
-package com.daniel.community.service;
+package com.daniel.community.post.service;
 
-import com.daniel.community.dto.CreatePostRequest;
-import com.daniel.community.dto.CreatePostResponse;
-import com.daniel.community.dto.PostDetailResponse;
-import com.daniel.community.dto.PostListResponse;
-import com.daniel.community.dto.PostSummaryResponse;
-import com.daniel.community.dto.UpdatePostRequest;
-import com.daniel.community.entity.Post;
-import com.daniel.community.repository.PostRepository;
+import com.daniel.community.post.dto.CreatePostRequest;
+import com.daniel.community.post.dto.CreatePostResponse;
+import com.daniel.community.post.dto.PostDetailResponse;
+import com.daniel.community.post.dto.PostListResponse;
+import com.daniel.community.post.dto.PostSummaryResponse;
+import com.daniel.community.post.dto.UpdatePostRequest;
+import com.daniel.community.post.entity.Post;
+import com.daniel.community.user.entity.User;
+import com.daniel.community.post.repository.PostRepository;
+import com.daniel.community.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
 
-// 서비스 계층
 @Service
 public class PostService {
-    // 인증 인가 구현 이전이므로 임의의 사용자로 대체
-    private static final String DEFAULT_NICKNAME = "startup";
-    private static final String DEFAULT_PROFILE_IMAGE = "https://image.kr/img.jpg";
 
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
-    // Service는 DB에 직접 접근하지 않고 Repository를 통해 접근
-    public PostService(PostRepository postRepository) {
+    // Repository를 통해 DB 접근
+    public PostService(
+            PostRepository postRepository,
+            UserRepository userRepository
+    ) {
         this.postRepository = postRepository;
+        this.userRepository = userRepository;
     }
 
     @Transactional
-    // createPostRequest 에서 값을 꺼내 Post 엔티티를 생성
-    public CreatePostResponse createPost(CreatePostRequest request) {
+    // createPostRequest에서 값을 꺼내 Post 엔티티를 생성
+    public CreatePostResponse createPost(CreatePostRequest request, Long userId) {
+        User user = findUser(userId);
+
         Post post = new Post(
                 request.getTitle(),
                 request.getContent(),
                 request.getPostImage(),
-                DEFAULT_NICKNAME,
-                DEFAULT_PROFILE_IMAGE
+                user
         );
 
         Post savedPost = postRepository.save(post);
+
         // 저장 이후에 생성된 게시물의 ID를 응답으로 반환
         return new CreatePostResponse(savedPost.getPostId());
     }
@@ -88,23 +93,41 @@ public class PostService {
 
     // 게시글 수정
     @Transactional
-    public void updatePost(Long postId, UpdatePostRequest request) {
+    public void updatePost(Long postId, UpdatePostRequest request, Long userId) {
         Post post = findPost(postId);
+
+        // 게시글의 작성자인지 확인
+        if (!post.isWrittenBy(userId)) {
+            throw new IllegalArgumentException("forbidden");
+        }
         // Entity의 update 메서드 호출
         post.update(request.getTitle(), request.getContent(), request.getPostImage());
     }
 
+    // 게시글 삭제
     @Transactional
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, Long userId) {
         Post post = findPost(postId);
+
+        // 게시글의 작성자인지 확인
+        if (!post.isWrittenBy(userId)) {
+            throw new IllegalArgumentException("forbidden");
+        }
         // Entity의 delete 메서드 호출
         postRepository.delete(post);
     }
 
-    // 게시글을 찾는 메서드
+    // 게시글 탐색
     private Post findPost(Long postId) {
         return postRepository.findById(postId)
                 // 게시글이 없을 경우 예외 처리
                 .orElseThrow(() -> new IllegalArgumentException("post_not_found"));
+    }
+
+    // 사용자 탐색
+    private User findUser(Long userId) {
+        return userRepository.findById(userId)
+                // 사용자가 없을 경우 예외 처리
+                .orElseThrow(() -> new IllegalArgumentException("unauthorized"));
     }
 }
